Move `RemoteFeedImageDataLoader` to production and change access control
